feat: вынесена логика чата в хук
